AniplannerV3/requirements.txt
flask
requests
openai
rapidfuzz
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv